modify fe_easiroc01

+++ Frontend/fe_easiroc01_node/script/copy_inputdac.sh
@@ -2,12 +2,34 @@
 
 bin_dir=$(cd $(dirname $0); pwd -P)
 
-param_dir=/home/DAQ/bftparam/201510_e0705/inputdac_g30_temp23.0c
+param_dir=$bin_dir/Input8bitDAC
+
+#temp_dir=$param_dir/inputdac_g30_temp19.0
+#temp_dir=$param_dir/inputdac_g30_temp20.0
+#temp_dir=$param_dir/inputdac_g30_temp21.0
+temp_dir=$param_dir/inputdac_g30_temp22.0
+#temp_dir=$param_dir/inputdac_g30_temp23.0
+#temp_dir=$param_dir/inputdac_g30_temp24.0
+#temp_dir=$param_dir/inputdac_g30_temp25.0
+
+log_file=$bin_dir/last_param.log
+date > $log_file
+echo "" >> $log_file
+echo "bin_dir   : $bin_dir"   >> $log_file
+echo "param_dir : $param_dir" >> $log_file
+echo "temp_dir  : $(basename $temp_dir)"  >> $log_file
 
 for i in $(seq 0 9)
   do
   nickname=easiroc`printf "%02d" $i`
-  param_file=$param_dir/Input8bitDAC_$i.txt
+  param_file=$temp_dir/Input8bitDAC_$i.txt
+  
+  if [ ! -f $param_file ]; then
+      echo -e "\e[35mcannot find parameter\e[m : $param_file"
+      echo -e "\e[35mcannot find parameter\e[m : $param_file" >> $log_file
+      continue;
+  fi
 
+  echo "$nickname : $(basename $param_file)" >> $log_file
   cp -v $param_file $bin_dir/$nickname/setup/Input8bitDAC.txt
 done
